--- apps/website/src/module/root/ui/footer.story.tsx
@@ -1,13 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { Footer } from './footer.layout';
-
-type Story = StoryObj<typeof Footer>;
-
-const meta = {
-  component: Footer,
-  argTypes: {},
-} satisfies Meta<typeof Footer>;
-
-export default meta;
-
-export const Default: Story = {};
