+++ apps/website/src/module/root/ui/layout/header/index.tsx
@@ -1,0 +1,1 @@
+export * from './header.presenter';
